+++ src/blas/arithmetic/add2others.sh
@@ -1,3 +1,3 @@
-cat add.cpp | sed -e 's; +; -;g' | sed -e 's;operator+;operator-;g'  > sub.cpp
-cat add.cpp | sed -e 's; +; *;g' | sed -e 's;operator+;operator*;g'  > mul.cpp
-cat add.cpp | sed -e 's; +; /;g' | sed -e 's;operator+;operator/;g'  > div.cpp
+cat add.cpp | sed -e 's/vadd/vsub/g' | sed -e 's;operator+;operator-;g' > sub.cpp
+cat add.cpp | sed -e 's/vadd/vmul/g' | sed -e 's;operator+;operator*;g' > mul.cpp
+cat add.cpp | sed -e 's/vadd/vdiv/g' | sed -e 's;operator+;operator/;g' > div.cpp