+++ test/matrix/matrix_arithmetic.cpp
@@ -1,4 +1,5 @@
 #include "arithmetic/mm_add.hpp"
+#include "arithmetic/mm_copy.hpp"
 #include "arithmetic/mm_mul.hpp"
 #include "arithmetic/mv_mul.hpp"
 
@@ -149,6 +150,48 @@ int main(int argc, char **argv) {
   if (test_mm_mul<monolish::matrix::CRS<float>, monolish::matrix::Dense<float>,
                   monolish::matrix::Dense<float>, float>(M, N, K, 1.0e-4) ==
       false) {
+    return 1;
+  }
+
+  // mm_copy Dense//
+  if (test_send_mm_copy<monolish::matrix::Dense<double>,
+                        monolish::matrix::Dense<double>, double>(
+          M, N, 1.0e-8) == false) {
+    return 1;
+  }
+  if (test_send_mm_copy<monolish::matrix::Dense<float>,
+                        monolish::matrix::Dense<float>, float>(M, N, 1.0e-4) ==
+      false) {
+    return 1;
+  }
+  if (test_mm_copy<monolish::matrix::Dense<double>,
+                   monolish::matrix::Dense<double>, double>(M, N, 1.0e-8) ==
+      false) {
+    return 1;
+  }
+  if (test_mm_copy<monolish::matrix::Dense<float>,
+                   monolish::matrix::Dense<float>, float>(M, N, 1.0e-4) ==
+      false) {
+    return 1;
+  }
+
+  // mm_copy CRS//
+  if (test_send_mm_copy<monolish::matrix::CRS<double>,
+                        monolish::matrix::CRS<double>, double>(M, N, 1.0e-8) ==
+      false) {
+    return 1;
+  }
+  if (test_send_mm_copy<monolish::matrix::CRS<float>,
+                        monolish::matrix::CRS<float>, float>(M, N, 1.0e-4) ==
+      false) {
+    return 1;
+  }
+  if (test_mm_copy<monolish::matrix::CRS<double>, monolish::matrix::CRS<double>,
+                   double>(M, N, 1.0e-8) == false) {
+    return 1;
+  }
+  if (test_mm_copy<monolish::matrix::CRS<float>, monolish::matrix::CRS<float>,
+                   float>(M, N, 1.0e-4) == false) {
     return 1;
   }